+++ sdks/python/hyperspace/__init__.py
@@ -1,3 +1,21 @@
 from .client import HyperspaceClient
+from .embedders import (
+    BaseEmbedder,
+    OpenAIEmbedder,
+    OpenRouterEmbedder,
+    CohereEmbedder,
+    VoyageEmbedder,
+    GoogleEmbedder,
+    SentenceTransformerEmbedder
+)
 
-__all__ = ["HyperspaceClient"]
+__all__ = [
+    "HyperspaceClient",
+    "BaseEmbedder",
+    "OpenAIEmbedder",
+    "OpenRouterEmbedder",
+    "CohereEmbedder",
+    "VoyageEmbedder",
+    "GoogleEmbedder",
+    "SentenceTransformerEmbedder"
+]